copernicoaches/dia1/pyproject.toml
[project]
name = "dia1"
version = "0.1.0"
description = "Dia 1 Flask app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
